producer/requirements.txt
kafka-python==2.0.2
faker==26.0.0
python-dateutil==2.9.0.post0
